refactor: moved layout model file and made renaming when needed
